refactor: small refactoring of file structure
